Changed variable names
Changed the name of the **target** from 'dev' (for device) to 'target'. This
is the term used by Cypress for the system being updated.

Changed the name of the firmware image file from 'prog' to 'fw_img'.
Cypress uses either the term **application** or **firmware image** for
the data that is placed in flash.