Implement re-scanning
The 2nd command line argument now works as intended. If not included or
connection fails, the scanner UI will start. Note that if a bad argument
is given, the program will halt with the usage statement.

Also, if the chosen device from the UI fails to connect, the UI will be
reset and devices will be re-scanned. This process will loop until the
user chooses a device that is successfully connected to.